+++ conf/soap.php
@@ -1,37 +1,14 @@
 <?php
 
 /**
- * Springy Framework Configuration File.
+ * Entradas:
+ *   - 'useCURL' - determina se o modo CURL deve ser usado.
+ *   - 'timeout' - determina o timeout em segundos.
  *
- * Do not remove this file.
- */
-
-/**
- *  \defgroup soapcfg Configurações da classe SOAP
- *  \ingroup config.
- *
- *  As entradas de configuração dos arquivos \c soap, são utilizadas pela classe SOAP, SOAP_Server e SOAP_Client,
- *  sendo que as entradas previamente definidas não podem ser omitidas sob risco dessas classes não funcionarem.
- *
- *  Você poderá adicionar entradas personalizadas de sua aplicação.
- *
- *  Entradas de configuração:
- *  - \c useCURL - determina se o modo CURL deve ser usado
- *  - \c timeout - determina o timeout em segundos
- *
- *  \see config
+ * @deprecated 4.5
  **/
 
-/**
- *  \defgroup soapcfg_default Configurações da classe SOAP para todos os ambientes
- *  \ingroup soapcfg.
- *
- *  As entradas colocadas nesse arquivo serão aplicadas a todos os ambientes do sistema.
- *
- *  Veja \link soapcfg Configurações da classe SOAP \endlink para entender as entradas de configuração possíveis.
- *
- *  \see soapcfg
- */
-/**@{*/
-
-return [];
+return [
+    'useCURL' => false,
+    'timeout' => 15,
+];